+++ lib/firstfit.c
@@ -1,10 +1,10 @@
 #include <firstfit.h>
 
-void *ff_alloc(void *freelist, unsigned long size)
+void *ff_alloc(void *freelist, size_t size)
 {
 	struct ff_freelist_t *p, *new, **head;
 	struct list_t *curr;
-	unsigned long size_min;
+	size_t size_min;
 
 	head = (struct ff_freelist_t **)freelist;
 	curr = &(*head)->list;
@@ -25,8 +25,8 @@ void *ff_alloc(void *freelist, unsigned long size)
 		if (size_min <= p->size) {
 			new = p;
 
-			p = (struct ff_freelist_t *)((unsigned long)p + size_min);
-			p->addr = (void *)((unsigned long)p +
+			p = (struct ff_freelist_t *)((unsigned int)p + size_min);
+			p->addr = (void *)((unsigned int)p +
 					sizeof(struct ff_freelist_t));
 			p->size = new->size - size_min;
 			list_add(&p->list, curr);
@@ -56,7 +56,7 @@ void ff_free(void *freelist, void *addr)
 	head = (struct ff_freelist_t **)freelist;
 	curr = &(*head)->list;
 	new  = (struct ff_freelist_t *)
-		((unsigned long)addr - sizeof(struct ff_freelist_t));
+		((unsigned int)addr - sizeof(struct ff_freelist_t));
 
 	do {
 		if (&new->list < curr) {
@@ -73,8 +73,8 @@ ff_merge:
 				p = get_container_of(curr,
 						struct ff_freelist_t, list);
 
-				if (((unsigned long)new->addr + new->size) ==
-						(unsigned long)p) {
+				if (((unsigned int)new->addr + new->size) ==
+						(unsigned int)p) {
 					new->size += p->size +
 						sizeof(struct ff_freelist_t);
 					list_del(curr);
@@ -105,9 +105,9 @@ struct ff_freelist_t *ff_freelist_init(void *start, void *end)
 	struct ff_freelist_t *head;
 
 	head = (struct ff_freelist_t *)ALIGN_WORD(start);
-	head->size = (unsigned long)end -
-		((unsigned long)head + sizeof(struct ff_freelist_t));
-	head->addr = (void *)((unsigned long)head + sizeof(struct ff_freelist_t));
+	head->size = (unsigned int)end -
+		((unsigned int)head + sizeof(struct ff_freelist_t));
+	head->addr = (void *)((unsigned int)head + sizeof(struct ff_freelist_t));
 	list_link_init(&head->list);
 
 	return head;
